Fix file read buffer allocation resizing size
I was using the FindData low size alone, rather than combining the high
and low sizes to get the true file size. This resulted in allocating a
buffer that was too small for really big files. The fread call did not
trigger any faults even when reading more data than room had been
allocated for. MeowHash then would trigger the access violation when it
tried to read memory outside of the allocated region.